refactor(markdown)!: the CST verbs are reached only through MD.cst
Moving the CST out of its sub-package put `Cst`, `CstParser` and `Lower` in
the root package, where `import morphir.langkit.markdown.*` reaches them. That
gave every CST operation two spellings — `Lower.lower(document)` and
`MD.cst.lower(document)` — and put three object names into the compatibility
surface that no caller should have to know.

The three join `Parser` and `MdWriter` in `internal`, `private[markdown]`, and
`MD.cst` re-exports their members as it always did. Nothing private escapes: a
generated forwarder's signature names only `MdCstNode.Document`, `String`,
`Chunk[String]` and `MdNode.Root`, all public. A probe outside the package
confirms the boundary rather than trusting it — `package morphir.langkit.probe`
naming `internal.Lower` does not compile.

`MdCstNode` and `LinkForm` stay public in `morphir.langkit.markdown`, which is
the line the module's rule actually draws: a type a caller's own signature has
to name must be reachable, an operation a caller only performs need not be.
`MdCstNode` splits into its own file to sit on the right side of that line.

This does not undo the one-public-package fix. That rule governs public types,
and machinery is what `internal` is for — the same reasoning that sent
`CstFragment` and `InlineNotes` there rather than keeping a `cst` package
alive for them.

CONTRIBUTING's table was stale in both directions and is now true: it named
`Parser` and `ParseError`, which are `internal.Parser` and `MdParseError`, and
it would otherwise have listed these three as public. It states the types-are-
public, verbs-are-not rule the table is an instance of.
