+++ morphir/langkit/markdown/test/src/morphir/langkit/markdown/CstBlockStructureTests.scala
@@ -2,6 +2,7 @@ package morphir.langkit.markdown
 
 import kyo.*
 import kyo.test.*
+import morphir.langkit.markdown.internal.CstParser
 
 /**
  * That graduated blocks come back typed, and ungraduated regions stay verbatim.
