+++ morphir/langkit/markdown/trees/test/src/morphir/langkit/markdown/trees/MarkdownTreesTests.scala
@@ -3,8 +3,8 @@ package morphir.langkit.markdown.trees
 import kyo.*
 import kyo.test.*
 import morphir.langkit.markdown.{MdNode, MdProfile}
-import morphir.langkit.markdown.internal.Parser
-import morphir.langkit.markdown.{Cst, MdCstNode, CstParser}
+import morphir.langkit.markdown.MdCstNode
+import morphir.langkit.markdown.internal.{CstParser, Parser}
 import morphir.langkit.trees.{NodeTypeName, QueryableTree}
 import morphir.langkit.trees.unist.UnistProjection
